# Explosives - Optimize creating explosive place actions