**When merged this pull request will:**
- Optimize explosive place action (remove useless condition) and creating its children (optimized function and return is cached)

All of these were tested with `count magazines = 28` and `placeable explosives => 6, 2, 1, 1` (where each number represents the amount of a unique type of explosive).
- The old function took ~0.31 ms
- Useless condition on `ACE_Place` action took ~0.1 ms (time it takes `hasExplosives`)
- New function takes ~0.20 ms (without any caching)
- New function first time (with caching) takes ~0.23 ms (cachedCall adds ~0.03 ms)
- New function (when return is cached) takes ~0.009 ms

**Overall**: Instead of ~0.41 ms every second, the new method is ~0.23 ms after a change in the player's loadout and then ~0.009 ms (accessing the cached value) every second after that.

For an inventory with no placeable explosives and `count magazines = 17`: it is ~0.11 ms vs ~0.10 ms (first time after loadout change) and ~0.009 ms every time after that.