@mharis001 2018-06-30
The `hasExplosives` function is not used anywhere else after this change. The header says that it is a public function but it is not on the wiki. Safe to remove it?