@TheMagnetar 2018-06-30
I would not remove a public function. Some people do look at the code and by removing it you may break somethinget like other mode or missions