@dedmen 2018-06-30
hasExplosives still sounds useful. But atleast add that arrayIntersect fix to it