@jonpas 2018-06-30
@TheMagnetar are you using it? If it's not marked "Public: Yes" it's not meant for public use. Of course if someone uses it still, it's up to them to pay attention and update it or request it to be made public API properly.