@alganthe 2018-07-01
^ hasExplosive is indeed marked public jonpas.

imho if public funcs are removed a core dev should make an announcement on the forums and website to have the maximum visibility.

As to the usefulness of said func, i'd say it has it's place.