@mharis001 2018-07-01
That use is taken out in this PR, if the player has no explosives, there wont be any children actions and the action wont show anyway. Will update `hasExplosives` to be more efficient since it not being removed.