@jonpas 2018-07-01
Oh maybe I should've read the other comments properly... Yeah, as a Public function we have to keep backwards compatibility. If it's essentially deprecated, then mark it as so with the `ACE_DEPRECATED` macro (+2 major versions from the one it was added in). No need to improve, only remove when the time comes! :D