@mharis001 2018-07-02
The function isn't getting a better alternative, I just thought about removing it since it isn't used anywhere is ace. Still deprecate?